requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework==3.9.3
celery==4.4.6
redis==3.5.3
gunicorn==20.0.4
python-dotenv==0.13.0